refactor: share RPC and configuration contracts
