+++ src/channels.ts
@@ -1,3 +1,1 @@
-export const MNEMON_READ_CHANNEL = '/dsh-mnemon-read'
-export const MNEMON_WRITE_CHANNEL = '/dsh-mnemon-write'
-export const MNEMON_PACK_CHANNEL = '/dsh-mnemon-pack'
+export { MNEMON_PACK_CHANNEL, MNEMON_READ_CHANNEL, MNEMON_WRITE_CHANNEL } from './shared/contracts.ts'
